+++ compile/compiler.go
@@ -145,6 +145,8 @@ func (c compiler) load(p string) (*Module, error) {
 		Types:      make(map[string]TypeSpec),
 		Services:   make(map[string]*ServiceSpec),
 	}
+
+	m.Raw = s
 	c.Modules[p] = m
 	// the module is added to the map before processing includes to break
 	// cyclic includes.