+++ gen/generate.go
@@ -70,6 +70,9 @@ type Options struct {
 
 	// Do not generate service helpers
 	NoServiceHelpers bool
+
+	// Do not embed IDLs in generated code
+	NoEmbedIDL bool
 }
 
 // Generate generates code based on the given options.
@@ -162,6 +165,10 @@ type thriftPackageImporter struct {
 // given Thrift file relative to the ImportPrefix.
 func (i thriftPackageImporter) RelativePackage(file string) (string, error) {
 	return filepath.Rel(i.ThriftRoot, strings.TrimSuffix(file, ".thrift"))
+}
+
+func (i thriftPackageImporter) RelativeThriftFilePath(file string) (string, error) {
+	return filepath.Rel(i.ThriftRoot, file)
 }
 
 // Package returns the import path for the top-level package of the given Thrift
@@ -265,6 +272,20 @@ func generateModule(m *compile.Module, i thriftPackageImporter, builder *generat
 		if !o.NoTypes {
 			files["types.go"] = buff.Bytes()
 		}
+	}
+
+	if !o.NoEmbedIDL {
+		if err := embedIDL(g, i, m); err != nil {
+			return nil, err
+		}
+
+		buff := new(bytes.Buffer)
+		if err := g.Write(buff, token.NewFileSet()); err != nil {
+			return nil, fmt.Errorf(
+				"could not generate idl.go for %q: %v", m.ThriftPath, err)
+		}
+
+		files["idl.go"] = buff.Bytes()
 	}
 
 	// Services must be generated last because names of user-defined types take